Refactor: Trim codebase to bare essentials
This major refactoring removes bloat and experimental code to create a cleaner, more maintainable codebase.

## Changes:

### Removed redundant implementations:
- Deleted 6 experimental backtest engines (keeping only production-ready engine.rs)
- Deleted 3 redundant market data readers (keeping only optimized reader.rs)
- Removed unused FeatureCollector module

### Reorganized code:
- Moved 11 binary performance test files to benchmarks/
- Created strategies/utils.rs for shared calculations
- Centralized tick size configuration

### Cleaned up documentation:
- Removed outdated performance analysis files
- Removed obsolete implementation progress docs

### Results:
- Reduced codebase by ~7,000 lines
- Eliminated duplicate implementations
- Improved maintainability
- Preserved all core functionality

The codebase now focuses on production-ready components while moving experimental code to appropriate directories.
