+++ benchmarks/backtest_performance_comparison.rs
