+++ benchmarks/optimized_backtest.rs
