+++ benchmarks/profile_all_events.rs
