+++ benchmarks/simple_backtest_performance.rs
