--- docs/FEATURE_ADDITIONS_SUMMARY.md
@@ -1,147 +1,0 @@
-# Feature Additions Summary
-
-## Overview
-
-I've successfully implemented the missing features identified in the feature analysis. All new modules are fully unit-tested and integrated with the existing feature extraction system.
-
-## New Feature Modules
-
-### 1. Time Features (`time_features.rs`)
-Provides time and session-aware features critical for trading strategies.
-
-**Key Features:**
-- **Trading Session Detection**: RTH, pre-market, after-hours, overnight
-- **Session Timing**: Time to close, time since open
-- **Volume Profiles**: Session volume percentiles, hourly tracking
-- **Contract Awareness**: Days to expiry, roll activity indicators
-- **Intraday Patterns**: Hour buckets, day of week
-
-**Example Usage:**
-```rust
-let mut time_features = TimeFeatures::new(SessionConfig::default());
-time_features.set_expiry(instrument_id, expiry_date);
-time_features.update_volume(instrument_id, quantity, timestamp);
-```
-
-### 2. Context Features (`context_features.rs`)
-Tracks strategy state and position-aware features.
-
-**Key Features:**
-- **Position Tracking**: Size, side, inventory deviation
-- **P&L Metrics**: Realized, unrealized, daily P&L, drawdowns
-- **Risk Utilization**: Position limits, loss limits, order rate limits
-- **Trading Activity**: Trade counts, fill rates, consecutive wins/losses
-- **Order Activity**: Market/limit orders, cancellation rates
-
-**Example Usage:**
-```rust
-let mut context = ContextFeatures::new(risk_limits);
-context.update_position(fill_price, fill_quantity, timestamp);
-context.update_order_sent(is_market, timestamp);
-```
-
-### 3. Feature Transformations (`transformations.rs`)
-Provides ML-ready feature preprocessing.
-
-**Key Components:**
-- **FeatureScaler**: Z-score, min-max, and robust scaling
-- **FeatureEngineer**: Polynomial features, interactions, log transforms
-- **FeatureValidator**: Data quality checks, missing value detection
-
-**Example Usage:**
-```rust
-// Scaling
-let mut scaler = FeatureScaler::new(ScalingMethod::ZScore);
-scaler.fit(&collector, instrument_id);
-scaler.transform(&mut features);
-
-// Engineering
-let engineer = FeatureEngineer::new()
-    .with_polynomial(2)
-    .with_interactions();
-engineer.transform(&mut features);
-
-// Validation
-let validator = FeatureValidator::new()
-    .with_required(vec!["critical_feature".to_string()]);
-let report = validator.validate(&collector, instrument_id);
-```
-
-## Integration with Existing System
-
-### Extended FeatureExtractor
-The main `FeatureExtractor` now includes:
-
-```rust
-pub struct FeatureExtractor {
-    // Existing features
-    book_features: HashMap<InstrumentId, BookFeatures>,
-    flow_features: HashMap<InstrumentId, FlowFeatures>,
-    rolling_features: HashMap<InstrumentId, RollingFeatures>,
-    
-    // New features
-    time_features: Option<TimeFeatures>,
-    context_features: HashMap<InstrumentId, ContextFeatures>,
-}
-```
-
-### Context-Aware Extraction
-New method for strategy-aware feature extraction:
-
-```rust
-pub fn extract_with_context(
-    &mut self,
-    instrument_id: InstrumentId,
-    timestamp: u64,
-    context: &StrategyContext,
-) -> FeatureVector
-```
-
-## Performance Considerations
-
-All new features maintain the high-performance standards:
-- **Event-driven updates**: No polling or unnecessary computation
-- **Efficient data structures**: Pre-allocated buffers, HashMap lookups
-- **Minimal allocations**: Reuse of buffers, in-place updates
-- **Cache-friendly**: Aligned structures where beneficial
-
-## Testing Coverage
-
-Each module includes comprehensive unit tests:
-- **Time Features**: 5 tests covering sessions, timing, volume profiles, expiry
-- **Context Features**: 6 tests covering positions, P&L, order tracking
-- **Transformations**: 6 tests covering scaling, engineering, validation
-
-Total: 38 feature tests, all passing
-
-## Usage in Backtesting
-
-The new features integrate seamlessly with the planned backtesting architecture:
-
-```rust
-// In strategy implementation
-fn on_market_event(&mut self, event: &MarketEvent, context: &StrategyContext) -> StrategyOutput {
-    // Extract features with full context
-    let features = self.feature_extractor.extract_with_context(
-        event.instrument_id(),
-        event.timestamp(),
-        context
-    );
-    
-    // Use features for decision making
-    let signal = self.generate_signal(&features);
-    
-    // Convert to orders
-    self.signal_to_orders(signal, context)
-}
-```
-
-## Next Steps
-
-With these features implemented, the system is ready for:
-1. Strategy trait implementation
-2. Backtesting engine development
-3. ML model integration
-4. Performance optimization
-
-The feature infrastructure now provides all necessary market microstructure, temporal, and contextual information needed for sophisticated trading strategies.
