--- src/backtest/engine_fast.rs
@@ -1,369 +1,0 @@
-//! High-performance backtesting engine
-//!
-//! This optimized version reduces overhead through:
-//! 1. Direct event processing (no priority queue)
-//! 2. Lockless market state updates
-//! 3. Batch processing for cache efficiency
-//! 4. Pre-allocated buffers
-//! 5. Inline hot paths
-
-use crate::backtest::events::FillEvent;
-use crate::backtest::execution::{ExecutionEngine, FillModel, LatencyModel};
-use crate::backtest::market_state::MarketStateManager;
-use crate::backtest::metrics::{MetricsCollector, PerformanceMetrics};
-use crate::backtest::position::PositionManager;
-use crate::core::MarketUpdate;
-use crate::core::traits::MarketDataSource;
-use crate::core::types::{InstrumentId, Price, Quantity};
-use crate::features::FeaturePosition;
-use crate::market_data::FileReader;
-use crate::market_data::events::{MarketEvent, TradeEvent};
-use crate::strategy::{Strategy, StrategyContext};
-use std::collections::HashMap;
-use std::path::{Path, PathBuf};
-use std::sync::{Arc, RwLock};
-
-/// Fast backtest configuration
-#[derive(Debug, Clone)]
-pub struct FastBacktestConfig {
-    pub latency_model: LatencyModel,
-    pub fill_model: FillModel,
-    pub commission_per_contract: f64,
-    pub initial_capital: f64,
-    pub batch_size: usize,
-}
-
-impl Default for FastBacktestConfig {
-    fn default() -> Self {
-        Self {
-            latency_model: LatencyModel::Fixed(100),
-            fill_model: FillModel::MidPoint,
-            commission_per_contract: 0.5,
-            initial_capital: 100_000.0,
-            batch_size: 1000, // Process in batches for cache efficiency
-        }
-    }
-}
-
-/// Lightweight strategy state
-struct FastStrategyState {
-    strategy: Box<dyn Strategy>,
-    context: StrategyContext,
-    position: HashMap<InstrumentId, i64>,
-    capital: f64,
-}
-
-/// High-performance backtesting engine
-pub struct FastBacktestEngine {
-    config: FastBacktestConfig,
-    strategies: HashMap<String, FastStrategyState>,
-    market_state: Arc<RwLock<MarketStateManager>>,
-    execution_engine: ExecutionEngine,
-    position_manager: PositionManager,
-    metrics_collector: MetricsCollector,
-    current_time: u64,
-    events_processed: usize,
-    // Pre-allocated buffers
-    event_buffer: Vec<MarketUpdate>,
-    order_buffer: Vec<crate::strategy::OrderRequest>,
-}
-
-impl FastBacktestEngine {
-    pub fn new(config: FastBacktestConfig) -> Self {
-        let market_state = Arc::new(RwLock::new(MarketStateManager::new()));
-
-        Self {
-            execution_engine: ExecutionEngine::new(
-                config.latency_model.clone(),
-                config.fill_model.clone(),
-                market_state.clone(),
-            ),
-            position_manager: PositionManager::new(crate::features::RiskLimits::default()),
-            metrics_collector: MetricsCollector::new(config.initial_capital),
-            market_state,
-            config,
-            strategies: HashMap::new(),
-            current_time: 0,
-            events_processed: 0,
-            event_buffer: Vec::with_capacity(10000),
-            order_buffer: Vec::with_capacity(100),
-        }
-    }
-
-    /// Add strategy (optimized)
-    pub fn add_strategy(&mut self, mut strategy: Box<dyn Strategy>) -> Result<(), String> {
-        let strategy_config = strategy.config();
-        let strategy_id = strategy_config.name.clone();
-        let risk_limits = crate::features::RiskLimits::default();
-
-        // Create context
-        let context = StrategyContext::new(
-            strategy_id.clone(),
-            self.current_time,
-            FeaturePosition::default(),
-            risk_limits.clone(),
-            true,
-        );
-
-        // Initialize strategy
-        strategy.initialize(&context)?;
-
-        // Create state
-        let state = FastStrategyState {
-            strategy,
-            context,
-            position: HashMap::new(),
-            capital: self.config.initial_capital,
-        };
-
-        // Register with position manager
-        self.position_manager
-            .add_strategy(strategy_id.clone(), risk_limits);
-
-        self.strategies.insert(strategy_id, state);
-        Ok(())
-    }
-
-    /// Run backtest with direct file processing
-    pub fn run<P: AsRef<Path>>(&mut self, data_files: &[P]) -> Result<FastEngineReport, String> {
-        let start_time = std::time::Instant::now();
-
-        // Process each file directly
-        for file_path in data_files {
-            self.process_file_direct(file_path)?;
-        }
-
-        let elapsed = start_time.elapsed();
-
-        // Generate report
-        Ok(FastEngineReport {
-            events_processed: self.events_processed,
-            elapsed_seconds: elapsed.as_secs_f64(),
-            throughput: self.events_processed as f64 / elapsed.as_secs_f64(),
-            performance_metrics: self.get_metrics(),
-        })
-    }
-
-    /// Process file directly without event queue
-    #[inline]
-    fn process_file_direct<P: AsRef<Path>>(&mut self, file_path: P) -> Result<(), String> {
-        let paths = vec![PathBuf::from(file_path.as_ref())];
-        let mut reader =
-            FileReader::new(paths).map_err(|e| format!("Failed to open file: {}", e))?;
-
-        // Clear buffers
-        self.event_buffer.clear();
-
-        // Process in batches
-        loop {
-            // Fill buffer
-            let mut batch_count = 0;
-            while batch_count < self.config.batch_size {
-                if let Some(update) = reader.next_update() {
-                    self.event_buffer.push(update);
-                    batch_count += 1;
-                } else {
-                    break;
-                }
-            }
-
-            if batch_count == 0 {
-                break; // No more events
-            }
-
-            // Process batch
-            self.process_batch()?;
-            self.event_buffer.clear();
-        }
-
-        Ok(())
-    }
-
-    /// Process a batch of events (optimized hot path)
-    #[inline]
-    fn process_batch(&mut self) -> Result<(), String> {
-        // Update market state for entire batch (single lock acquisition)
-        {
-            let mut market_state = self.market_state.write().unwrap();
-            for update in &self.event_buffer {
-                self.update_market_state_fast(&mut market_state, update);
-            }
-        }
-
-        // Process each event through strategies
-        let event_buffer_len = self.event_buffer.len();
-        for i in 0..event_buffer_len {
-            let update = &self.event_buffer[i];
-            self.current_time = self.get_update_timestamp(update);
-            self.events_processed += 1;
-
-            // Convert to market event
-            let market_event = self.convert_update_fast(update);
-
-            // Clear order buffer
-            self.order_buffer.clear();
-
-            // Process strategies inline (no intermediate collections)
-            for (_id, state) in &mut self.strategies {
-                // Update context time
-                state.context.current_time = self.current_time;
-
-                // Call strategy
-                let output = state
-                    .strategy
-                    .on_market_event(&market_event, &state.context);
-
-                // Process output inline
-                if !output.orders.is_empty() {
-                    for order in &output.orders {
-                        self.order_buffer.push(order.clone());
-                    }
-                }
-            }
-
-            // Submit orders to execution engine
-            for order in &self.order_buffer {
-                self.execution_engine.submit_order(
-                    order.clone(),
-                    "fast_strategy".to_string(),
-                    self.current_time,
-                );
-            }
-
-            // Process any immediate fills
-            let fills = self.execution_engine.process_orders(self.current_time);
-            for fill in fills {
-                self.process_fill_fast(fill)?;
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Fast market state update (no event conversion)
-    #[inline]
-    fn update_market_state_fast(
-        &self,
-        _market_state: &mut MarketStateManager,
-        update: &MarketUpdate,
-    ) {
-        // Direct update without intermediate conversion
-        match update {
-            MarketUpdate::Trade(_trade) => {
-                // Simplified - just track last trade
-                // In real implementation, update order books
-            }
-            MarketUpdate::OrderBook(_book_update) => {
-                // Update order book state
-            }
-        }
-    }
-
-    /// Fast event conversion
-    #[inline]
-    fn convert_update_fast(&self, update: &MarketUpdate) -> MarketEvent {
-        match update {
-            MarketUpdate::Trade(trade) => MarketEvent::Trade(TradeEvent {
-                instrument_id: trade.instrument_id,
-                trade_id: 0,
-                price: trade.price,
-                quantity: trade.quantity,
-                aggressor_side: trade.side,
-                timestamp: trade.timestamp,
-                buyer_order_id: None,
-                seller_order_id: None,
-            }),
-            MarketUpdate::OrderBook(_) => {
-                // Simplified conversion
-                MarketEvent::Trade(TradeEvent {
-                    instrument_id: 1,
-                    trade_id: 0,
-                    price: Price::from(100i64),
-                    quantity: Quantity::from(1u32),
-                    aggressor_side: crate::core::Side::Bid,
-                    timestamp: self.current_time,
-                    buyer_order_id: None,
-                    seller_order_id: None,
-                })
-            }
-        }
-    }
-
-    /// Fast fill processing
-    #[inline]
-    fn process_fill_fast(&mut self, fill: FillEvent) -> Result<(), String> {
-        // Update position manager
-        self.position_manager.apply_fill(&fill)?;
-
-        // Update metrics
-        self.metrics_collector.process_fill(&fill);
-
-        // Update strategy state (simplified)
-        if let Some(state) = self.strategies.get_mut(&fill.strategy_id) {
-            let position = state.position.entry(fill.instrument_id).or_insert(0);
-            match fill.side {
-                crate::core::Side::Bid => *position += fill.quantity.as_i64(),
-                crate::core::Side::Ask => *position -= fill.quantity.as_i64(),
-            }
-
-            state.capital -= fill.total_cost();
-            state.context.position.quantity = *position;
-
-            // Notify strategy
-            state.strategy.on_fill(
-                fill.price,
-                fill.quantity.as_i64(),
-                fill.timestamp,
-                &state.context,
-            );
-        }
-
-        Ok(())
-    }
-
-    /// Get timestamp from market update
-    #[inline]
-    fn get_update_timestamp(&self, update: &MarketUpdate) -> u64 {
-        match update {
-            MarketUpdate::Trade(trade) => trade.timestamp,
-            MarketUpdate::OrderBook(book) => book.timestamp,
-        }
-    }
-
-    /// Get performance metrics
-    pub fn get_metrics(&self) -> PerformanceMetrics {
-        // Simplified for now - return default metrics
-        PerformanceMetrics::default()
-    }
-}
-
-/// Fast engine report
-#[derive(Debug)]
-pub struct FastEngineReport {
-    pub events_processed: usize,
-    pub elapsed_seconds: f64,
-    pub throughput: f64,
-    pub performance_metrics: PerformanceMetrics,
-}
-
-impl FastEngineReport {
-    pub fn print_summary(&self) {
-        println!("\n=== Fast Backtest Results ===");
-        println!("Events processed: {}", self.events_processed);
-        println!("Time elapsed: {:.2}s", self.elapsed_seconds);
-        println!("Throughput: {:.0} events/s", self.throughput);
-        println!(
-            "Efficiency vs 18M target: {:.1}%",
-            (self.throughput / 18_000_000.0) * 100.0
-        );
-
-        println!("\nPerformance Metrics:");
-        println!("Total trades: {}", self.performance_metrics.total_trades);
-        println!(
-            "Win rate: {:.1}%",
-            self.performance_metrics.win_rate * 100.0
-        );
-        println!("Sharpe ratio: {:.2}", self.performance_metrics.sharpe_ratio);
-        println!("Total P&L: ${:.2}", self.performance_metrics.total_pnl);
-    }
-}
