--- src/backtest/engine_simple_fast.rs
@@ -1,238 +1,0 @@
-//! Simplified high-performance backtesting engine
-//!
-//! Key optimizations:
-//! 1. Direct file processing (no event queue)
-//! 2. Reduced allocations
-//! 3. Batch processing for cache efficiency
-
-use crate::backtest::market_state::MarketStateManager;
-use crate::backtest::metrics::MetricsCollector;
-use crate::backtest::position::PositionManager;
-use crate::backtest::{BacktestConfig, EngineReport, PerformanceMetrics};
-use crate::core::MarketUpdate;
-use crate::core::traits::MarketDataSource;
-use crate::features::{FeaturePosition, RiskLimits};
-use crate::market_data::FileReader;
-use crate::market_data::events::{MarketEvent, TradeEvent};
-use crate::strategy::{Strategy, StrategyContext};
-use std::path::{Path, PathBuf};
-use std::sync::{Arc, RwLock};
-
-const BATCH_SIZE: usize = 1000;
-
-/// Simple fast backtesting engine
-pub struct SimpleFastEngine {
-    config: BacktestConfig,
-    strategies: Vec<(Box<dyn Strategy>, StrategyContext)>,
-    market_state: Arc<RwLock<MarketStateManager>>,
-    _position_manager: PositionManager,
-    _metrics_collector: MetricsCollector,
-    current_time: u64,
-    events_processed: usize,
-    // Pre-allocated buffers
-    event_buffer: Vec<MarketUpdate>,
-}
-
-impl SimpleFastEngine {
-    pub fn new(config: BacktestConfig) -> Self {
-        let market_state = Arc::new(RwLock::new(MarketStateManager::new()));
-
-        Self {
-            market_state: market_state.clone(),
-            _position_manager: PositionManager::new(RiskLimits::default()),
-            _metrics_collector: MetricsCollector::new(config.initial_capital),
-            config,
-            strategies: Vec::new(),
-            current_time: 0,
-            events_processed: 0,
-            event_buffer: Vec::with_capacity(BATCH_SIZE * 2),
-        }
-    }
-
-    /// Add strategy
-    pub fn add_strategy(&mut self, mut strategy: Box<dyn Strategy>) -> Result<(), String> {
-        let strategy_config = strategy.config();
-        let strategy_id = strategy_config.name.clone();
-
-        // Create context
-        let context = StrategyContext::new(
-            strategy_id.clone(),
-            self.current_time,
-            FeaturePosition::default(),
-            RiskLimits::default(),
-            true,
-        );
-
-        // Initialize strategy
-        strategy.initialize(&context)?;
-
-        // Add to list
-        self.strategies.push((strategy, context));
-
-        Ok(())
-    }
-
-    /// Run backtest with optimized processing
-    pub fn run<P: AsRef<Path>>(&mut self, data_files: &[P]) -> Result<EngineReport, String> {
-        let start_time = std::time::Instant::now();
-
-        // Process each file
-        for file_path in data_files {
-            self.process_file_fast(file_path)?;
-        }
-
-        let _elapsed = start_time.elapsed();
-
-        // Generate simple report
-        Ok(EngineReport {
-            config: self.config.clone(),
-            events_processed: self.events_processed,
-            strategy_results: Vec::new(),
-            portfolio_stats: crate::backtest::position::PortfolioStats::default(),
-            performance_metrics: PerformanceMetrics::default(),
-            trades: Vec::new(),
-            equity_curve: Vec::new(),
-        })
-    }
-
-    /// Fast file processing
-    fn process_file_fast<P: AsRef<Path>>(&mut self, file_path: P) -> Result<(), String> {
-        println!("Processing file: {:?}", file_path.as_ref());
-
-        let paths = vec![PathBuf::from(file_path.as_ref())];
-        let mut reader =
-            FileReader::new(paths).map_err(|e| format!("Failed to open file: {}", e))?;
-
-        // Clear buffer
-        self.event_buffer.clear();
-
-        // Process in batches
-        loop {
-            // Fill buffer
-            let batch_start = self.event_buffer.len();
-            for _ in 0..BATCH_SIZE {
-                if let Some(update) = reader.next_update() {
-                    self.event_buffer.push(update);
-                } else {
-                    break;
-                }
-            }
-
-            let batch_size = self.event_buffer.len() - batch_start;
-            if batch_size == 0 {
-                break; // No more events
-            }
-
-            // Process batch
-            self.process_batch_fast(batch_start, batch_start + batch_size)?;
-        }
-
-        // Clear buffer after processing
-        self.event_buffer.clear();
-
-        Ok(())
-    }
-
-    /// Process a batch of events (hot path)
-    #[inline]
-    fn process_batch_fast(&mut self, start: usize, end: usize) -> Result<(), String> {
-        // Update market state for batch (single lock)
-        {
-            let _market_state = self.market_state.write().unwrap();
-            for i in start..end {
-                let update = &self.event_buffer[i];
-                // Simple market state update
-                match update {
-                    MarketUpdate::Trade(trade) => {
-                        self.current_time = trade.timestamp;
-                    }
-                    MarketUpdate::OrderBook(book) => {
-                        self.current_time = book.timestamp;
-                    }
-                }
-            }
-        }
-
-        // Process each event through strategies
-        for i in start..end {
-            self.events_processed += 1;
-            let update = &self.event_buffer[i];
-
-            // Convert to market event
-            let market_event = self.convert_update(update);
-
-            // Update all strategies
-            for (strategy, context) in &mut self.strategies {
-                // Update context time
-                context.current_time = self.current_time;
-
-                // Call strategy
-                let output = strategy.on_market_event(&market_event, context);
-
-                // Process any orders (simplified - just count)
-                if !output.orders.is_empty() {
-                    // In real implementation, would submit orders
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Convert market update to event
-    #[inline]
-    fn convert_update(&self, update: &MarketUpdate) -> MarketEvent {
-        match update {
-            MarketUpdate::Trade(trade) => MarketEvent::Trade(TradeEvent {
-                instrument_id: trade.instrument_id,
-                trade_id: 0,
-                price: trade.price,
-                quantity: trade.quantity,
-                aggressor_side: trade.side,
-                timestamp: trade.timestamp,
-                buyer_order_id: None,
-                seller_order_id: None,
-            }),
-            MarketUpdate::OrderBook(book) => {
-                // Simplified - convert to trade event
-                MarketEvent::Trade(TradeEvent {
-                    instrument_id: book.instrument_id,
-                    trade_id: 0,
-                    price: crate::core::types::Price::from(100i64),
-                    quantity: crate::core::types::Quantity::from(1u32),
-                    aggressor_side: crate::core::Side::Bid,
-                    timestamp: book.timestamp,
-                    buyer_order_id: None,
-                    seller_order_id: None,
-                })
-            }
-        }
-    }
-}
-
-/// Print performance comparison
-pub fn print_performance_comparison(original_throughput: f64, fast_throughput: f64) {
-    println!("\n=== PERFORMANCE COMPARISON ===");
-    println!("Original engine: {:.0} events/s", original_throughput);
-    println!("Fast engine: {:.0} events/s", fast_throughput);
-
-    let improvement = fast_throughput / original_throughput;
-    println!("Improvement: {:.1}x", improvement);
-
-    let original_efficiency = (original_throughput / 18_000_000.0) * 100.0;
-    let fast_efficiency = (fast_throughput / 18_000_000.0) * 100.0;
-
-    println!("\nEfficiency vs 18M target:");
-    println!("Original: {:.1}%", original_efficiency);
-    println!("Fast: {:.1}%", fast_efficiency);
-
-    if fast_throughput >= 15_000_000.0 {
-        println!("\n✅ SUCCESS: Achieved target performance!");
-    } else if fast_throughput >= 10_000_000.0 {
-        println!("\n✅ GOOD: Significant improvement");
-    } else if fast_throughput >= 6_000_000.0 {
-        println!("\n⚠️  MODERATE: Some improvement");
-    } else {
-        println!("\n❌ More optimization needed");
-    }
-}
