--- src/market_data/reader_optimized.rs
@@ -1,309 +1,0 @@
-//! Optimized zero-copy market data reader
-//!
-//! This version eliminates the major performance bottlenecks:
-//! 1. Removes MboMsg cloning (2-3x speedup)
-//! 2. Uses direct processing pipeline
-//! 3. Minimizes allocations in hot paths
-
-use crate::core::{MarketDataSource, MarketUpdate, Result, Trade, BBO};
-use crate::core::types::{InstrumentId, Price, Quantity};
-use crate::core::Side;
-use crossbeam_channel::{bounded, Receiver, Sender};
-use databento::dbn::{
-    decode::{DbnDecoder, DecodeRecord},
-    record::MboMsg,
-    enums::{Action, Side as DbnSide},
-};
-use memmap2::Mmap;
-use std::{
-    fs::File,
-    io::BufReader,
-    path::PathBuf,
-    thread,
-};
-
-const BATCH_SIZE: usize = 4 * 1024;
-
-/// Zero-copy market data reader
-pub struct OptimizedFileReader {
-    rx: Receiver<Option<Vec<MarketUpdate>>>,
-    _handle: thread::JoinHandle<()>,
-    current_batch: Vec<MarketUpdate>,
-    batch_index: usize,
-}
-
-impl OptimizedFileReader {
-    /// Create a new optimized file reader
-    pub fn new(paths: Vec<PathBuf>) -> Result<Self> {
-        let (tx, rx) = bounded(10);
-        
-        let handle = thread::spawn(move || {
-            Self::producer_thread(paths, tx);
-        });
-
-        Ok(Self {
-            rx,
-            _handle: handle,
-            current_batch: Vec::new(),
-            batch_index: 0,
-        })
-    }
-
-    /// Producer thread with zero-copy processing
-    fn producer_thread(paths: Vec<PathBuf>, tx: Sender<Option<Vec<MarketUpdate>>>) {
-        let mut batch = Vec::with_capacity(BATCH_SIZE);
-
-        for path in paths {
-            if let Err(e) = Self::process_file_optimized(&path, &tx, &mut batch) {
-                eprintln!("Error processing file {:?}: {}", path, e);
-            }
-        }
-
-        if !batch.is_empty() {
-            tx.send(Some(batch)).ok();
-        }
-        
-        tx.send(None).ok();
-    }
-
-    /// Process file with zero-copy optimization
-    fn process_file_optimized(
-        path: &PathBuf,
-        tx: &Sender<Option<Vec<MarketUpdate>>>,
-        batch: &mut Vec<MarketUpdate>,
-    ) -> Result<()> {
-        let file = File::open(path)?;
-        let mmap = unsafe { Mmap::map(&file)? };
-        let reader = BufReader::new(std::io::Cursor::new(&mmap[..]));
-        let mut decoder = DbnDecoder::with_zstd_buffer(reader)?;
-
-        // Pre-allocate conversion buffer to avoid allocations
-        let mut conversion_buffer = Vec::with_capacity(BATCH_SIZE);
-
-        while let Some(rec) = decoder.decode_record::<MboMsg>()? {
-            // Convert directly to MarketUpdate without cloning MboMsg
-            if let Some(update) = Self::convert_mbo_direct(&rec) {
-                conversion_buffer.push(update);
-                
-                if conversion_buffer.len() == BATCH_SIZE {
-                    // Efficiently swap buffers instead of cloning
-                    batch.extend(conversion_buffer.drain(..));
-                    tx.send(Some(std::mem::replace(batch, Vec::with_capacity(BATCH_SIZE))))?;
-                    conversion_buffer.clear();
-                }
-            }
-        }
-
-        // Handle remaining messages
-        if !conversion_buffer.is_empty() {
-            batch.extend(conversion_buffer.drain(..));
-        }
-
-        Ok(())
-    }
-
-    /// Direct conversion from MboMsg to MarketUpdate (zero-copy)
-    #[inline(always)]
-    fn convert_mbo_direct(mbo: &MboMsg) -> Option<MarketUpdate> {
-        let instrument_id = mbo.hd.instrument_id as InstrumentId;
-        let timestamp = mbo.ts_recv;
-        
-        match mbo.action {
-            Action::Add | Action::Modify => {
-                // Create BBO update for adds/modifies
-                let price = Price::new(mbo.price);
-                let quantity = Quantity::from(mbo.size as u32);
-                let side = convert_dbn_side(mbo.side);
-                
-                match side {
-                    Side::Bid => Some(MarketUpdate::BBO(BBO {
-                        instrument_id,
-                        bid_price: price,
-                        ask_price: Price::new(0), // Will be filled by book reconstruction
-                        bid_quantity: quantity,
-                        ask_quantity: Quantity::from(0u32),
-                        timestamp,
-                    })),
-                    Side::Ask => Some(MarketUpdate::BBO(BBO {
-                        instrument_id,
-                        bid_price: Price::new(0),
-                        ask_price: price,
-                        bid_quantity: Quantity::from(0u32),
-                        ask_quantity: quantity,
-                        timestamp,
-                    })),
-                }
-            }
-            Action::Trade => {
-                // Create trade update
-                Some(MarketUpdate::Trade(Trade {
-                    instrument_id,
-                    price: Price::new(mbo.price),
-                    quantity: Quantity::from(mbo.size as u32),
-                    side: convert_dbn_side(mbo.side),
-                    timestamp,
-                }))
-            }
-            _ => None, // Skip other actions for now
-        }
-    }
-}
-
-/// Convert databento side to our side enum (inline for performance)
-#[inline(always)]
-fn convert_dbn_side(side: DbnSide) -> Side {
-    match side {
-        DbnSide::Bid => Side::Bid,
-        DbnSide::Ask => Side::Ask,
-        DbnSide::None => Side::Bid, // Default
-    }
-}
-
-impl MarketDataSource for OptimizedFileReader {
-    fn next_update(&mut self) -> Option<MarketUpdate> {
-        // First, try to get from current batch
-        if self.batch_index < self.current_batch.len() {
-            let update = self.current_batch[self.batch_index].clone();
-            self.batch_index += 1;
-            return Some(update);
-        }
-
-        // Try to get next batch
-        match self.rx.recv() {
-            Ok(Some(batch)) => {
-                self.current_batch = batch;
-                self.batch_index = 0;
-                
-                if !self.current_batch.is_empty() {
-                    let update = self.current_batch[0].clone();
-                    self.batch_index = 1;
-                    Some(update)
-                } else {
-                    None
-                }
-            }
-            _ => None,
-        }
-    }
-}
-
-/// Lockless market state snapshot for strategy contexts
-#[derive(Debug, Clone)]
-pub struct MarketSnapshot {
-    /// Best bid/offer by instrument (pre-computed)
-    pub bbo: hashbrown::HashMap<InstrumentId, (Price, Price, Quantity, Quantity)>,
-    /// Last trade prices
-    pub last_prices: hashbrown::HashMap<InstrumentId, Price>,
-    /// Timestamps of last updates
-    pub last_updates: hashbrown::HashMap<InstrumentId, u64>,
-}
-
-impl MarketSnapshot {
-    pub fn new() -> Self {
-        Self {
-            bbo: hashbrown::HashMap::new(),
-            last_prices: hashbrown::HashMap::new(),
-            last_updates: hashbrown::HashMap::new(),
-        }
-    }
-    
-    /// Get BBO without locks (O(1) access)
-    #[inline(always)]
-    pub fn get_bbo(&self, instrument_id: InstrumentId) -> Option<(Price, Price)> {
-        self.bbo.get(&instrument_id).map(|(bid, ask, _, _)| (*bid, *ask))
-    }
-    
-    /// Get last price without locks
-    #[inline(always)]
-    pub fn get_last_price(&self, instrument_id: InstrumentId) -> Option<Price> {
-        self.last_prices.get(&instrument_id).copied()
-    }
-    
-    /// Update BBO efficiently
-    #[inline(always)]
-    pub fn update_bbo(&mut self, instrument_id: InstrumentId, bid: Price, ask: Price, 
-                      bid_qty: Quantity, ask_qty: Quantity, timestamp: u64) {
-        self.bbo.insert(instrument_id, (bid, ask, bid_qty, ask_qty));
-        self.last_updates.insert(instrument_id, timestamp);
-    }
-    
-    /// Update last price efficiently  
-    #[inline(always)]
-    pub fn update_last_price(&mut self, instrument_id: InstrumentId, price: Price, timestamp: u64) {
-        self.last_prices.insert(instrument_id, price);
-        self.last_updates.insert(instrument_id, timestamp);
-    }
-}
-
-/// Pre-allocated feature vector using indices instead of strings
-#[derive(Debug)]
-pub struct IndexedFeatureVector {
-    /// Feature values indexed by pre-defined constants
-    pub values: Vec<f64>,
-    /// Timestamps
-    pub timestamp: u64,
-}
-
-impl IndexedFeatureVector {
-    /// Create with pre-allocated size
-    pub fn new(size: usize, timestamp: u64) -> Self {
-        Self {
-            values: vec![0.0; size],
-            timestamp,
-        }
-    }
-    
-    /// Set feature by index (no string operations)
-    #[inline(always)]
-    pub fn set(&mut self, index: usize, value: f64) {
-        if index < self.values.len() {
-            self.values[index] = value;
-        }
-    }
-    
-    /// Get feature by index
-    #[inline(always)]
-    pub fn get(&self, index: usize) -> f64 {
-        self.values.get(index).copied().unwrap_or(0.0)
-    }
-}
-
-// Pre-defined feature indices (compile-time constants)
-pub mod feature_indices {
-    pub const SPREAD_ABSOLUTE: usize = 0;
-    pub const SPREAD_RELATIVE: usize = 1;
-    pub const BID_SIZE: usize = 2;
-    pub const ASK_SIZE: usize = 3;
-    pub const VOLUME_IMBALANCE: usize = 4;
-    pub const MIDPOINT_RETURN: usize = 5;
-    pub const VOLATILITY: usize = 6;
-    pub const TOTAL_FEATURES: usize = 7;
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    
-    #[test]
-    fn test_indexed_features() {
-        let mut features = IndexedFeatureVector::new(feature_indices::TOTAL_FEATURES, 1000);
-        
-        features.set(feature_indices::SPREAD_ABSOLUTE, 0.05);
-        features.set(feature_indices::BID_SIZE, 1000.0);
-        
-        assert_eq!(features.get(feature_indices::SPREAD_ABSOLUTE), 0.05);
-        assert_eq!(features.get(feature_indices::BID_SIZE), 1000.0);
-    }
-    
-    #[test]
-    fn test_market_snapshot() {
-        let mut snapshot = MarketSnapshot::new();
-        
-        snapshot.update_bbo(1, Price::new(100_000), Price::new(100_100), 
-                           Quantity::from(1000u32), Quantity::from(1200u32), 1000);
-        
-        let bbo = snapshot.get_bbo(1).unwrap();
-        assert_eq!(bbo.0, Price::new(100_000));
-        assert_eq!(bbo.1, Price::new(100_100));
-    }
-}
