--- src/market_data/reader_optimized_fixed.rs
@@ -1,357 +1,0 @@
-//! High-performance zero-copy market data reader
-//!
-//! This version eliminates major performance bottlenecks:
-//! 1. Removes MboMsg cloning (2-3x speedup)
-//! 2. Direct conversion to core types
-//! 3. Batch processing optimization
-
-use crate::core::{MarketDataSource, MarketUpdate, Result, Trade, BookUpdate, BookUpdateType};
-use crate::core::types::{InstrumentId, Price, Quantity, OrderId};
-use crate::core::Side;
-use crossbeam_channel::{bounded, Receiver, Sender};
-use databento::dbn::{
-    decode::{DbnDecoder, DecodeRecord},
-    record::MboMsg,
-};
-use memmap2::Mmap;
-use std::{
-    fs::File,
-    io::BufReader,
-    path::PathBuf,
-    thread,
-};
-
-const BATCH_SIZE: usize = 4 * 1024;
-
-/// Zero-copy optimized file reader
-pub struct OptimizedFileReader {
-    rx: Receiver<Option<Vec<MarketUpdate>>>,
-    _handle: thread::JoinHandle<()>,
-    current_batch: Vec<MarketUpdate>,
-    batch_index: usize,
-}
-
-impl OptimizedFileReader {
-    /// Create new optimized reader
-    pub fn new(paths: Vec<PathBuf>) -> Result<Self> {
-        let (tx, rx) = bounded(10);
-        
-        let handle = thread::spawn(move || {
-            Self::producer_thread(paths, tx);
-        });
-
-        Ok(Self {
-            rx,
-            _handle: handle,
-            current_batch: Vec::new(),
-            batch_index: 0,
-        })
-    }
-
-    /// Producer thread with optimized processing
-    fn producer_thread(paths: Vec<PathBuf>, tx: Sender<Option<Vec<MarketUpdate>>>) {
-        let mut batch = Vec::with_capacity(BATCH_SIZE);
-
-        for path in paths {
-            if let Err(e) = Self::process_file_optimized(&path, &tx, &mut batch) {
-                eprintln!("Error processing file {:?}: {}", path, e);
-            }
-        }
-
-        if !batch.is_empty() {
-            tx.send(Some(batch)).ok();
-        }
-        
-        tx.send(None).ok();
-    }
-
-    /// Process file with zero-copy optimization
-    fn process_file_optimized(
-        path: &PathBuf,
-        tx: &Sender<Option<Vec<MarketUpdate>>>,
-        batch: &mut Vec<MarketUpdate>,
-    ) -> Result<()> {
-        let file = File::open(path)?;
-        let mmap = unsafe { Mmap::map(&file)? };
-        let reader = BufReader::new(std::io::Cursor::new(&mmap[..]));
-        let mut decoder = DbnDecoder::with_zstd_buffer(reader)?;
-
-        // Pre-allocate conversion buffer
-        let mut conversion_buffer = Vec::with_capacity(BATCH_SIZE);
-
-        while let Some(rec) = decoder.decode_record::<MboMsg>()? {
-            // Convert directly without cloning MboMsg
-            if let Some(update) = Self::convert_mbo_zero_copy(&rec) {
-                conversion_buffer.push(update);
-                
-                if conversion_buffer.len() == BATCH_SIZE {
-                    batch.extend(conversion_buffer.drain(..));
-                    tx.send(Some(std::mem::replace(batch, Vec::with_capacity(BATCH_SIZE))))?;
-                    conversion_buffer.clear();
-                }
-            }
-        }
-
-        if !conversion_buffer.is_empty() {
-            batch.extend(conversion_buffer.drain(..));
-        }
-
-        Ok(())
-    }
-
-    /// Zero-copy conversion from MboMsg to MarketUpdate
-    #[inline(always)]
-    fn convert_mbo_zero_copy(mbo: &MboMsg) -> Option<MarketUpdate> {
-        let instrument_id = mbo.hd.instrument_id as InstrumentId;
-        let timestamp = mbo.ts_recv;
-        let order_id = mbo.order_id as OrderId;
-        
-        // Convert action type (i8) to our enums
-        match mbo.action {
-            1 => { // Add
-                let price = Price::new(mbo.price);
-                let quantity = Quantity::from(mbo.size as u32);
-                let side = convert_side_i8(mbo.side);
-                
-                Some(MarketUpdate::OrderBook(BookUpdate {
-                    instrument_id,
-                    update_type: BookUpdateType::Add {
-                        order_id,
-                        side,
-                        price,
-                        quantity,
-                    },
-                    timestamp,
-                }))
-            }
-            2 => { // Cancel
-                Some(MarketUpdate::OrderBook(BookUpdate {
-                    instrument_id,
-                    update_type: BookUpdateType::Cancel { order_id },
-                    timestamp,
-                }))
-            }
-            3 => { // Modify
-                let new_quantity = Quantity::from(mbo.size as u32);
-                Some(MarketUpdate::OrderBook(BookUpdate {
-                    instrument_id,
-                    update_type: BookUpdateType::Modify {
-                        order_id,
-                        new_quantity,
-                    },
-                    timestamp,
-                }))
-            }
-            4 => { // Trade
-                let price = Price::new(mbo.price);
-                let quantity = Quantity::from(mbo.size as u32);
-                let side = convert_side_i8(mbo.side);
-                
-                Some(MarketUpdate::Trade(Trade {
-                    instrument_id,
-                    price,
-                    quantity,
-                    side,
-                    timestamp,
-                }))
-            }
-            _ => None, // Skip unknown actions
-        }
-    }
-}
-
-/// Convert databento side (i8) to our Side enum (inline for performance)
-#[inline(always)]
-fn convert_side_i8(side: i8) -> Side {
-    match side {
-        1 => Side::Bid,
-        2 => Side::Ask,
-        _ => Side::Bid, // Default fallback
-    }
-}
-
-impl MarketDataSource for OptimizedFileReader {
-    fn next_update(&mut self) -> Option<MarketUpdate> {
-        // Get from current batch first
-        if self.batch_index < self.current_batch.len() {
-            let update = self.current_batch[self.batch_index].clone();
-            self.batch_index += 1;
-            return Some(update);
-        }
-
-        // Get next batch
-        match self.rx.recv() {
-            Ok(Some(batch)) => {
-                self.current_batch = batch;
-                self.batch_index = 0;
-                
-                if !self.current_batch.is_empty() {
-                    let update = self.current_batch[0].clone();
-                    self.batch_index = 1;
-                    Some(update)
-                } else {
-                    None
-                }
-            }
-            _ => None,
-        }
-    }
-    
-    fn subscribe(&mut self, _instruments: Vec<InstrumentId>) -> Result<()> {
-        // For file-based readers, subscription is not applicable
-        Ok(())
-    }
-}
-
-/// High-performance market state cache (lockless)
-#[derive(Debug, Clone)]
-pub struct FastMarketState {
-    /// Best prices by instrument (bid, ask)
-    pub best_prices: hashbrown::HashMap<InstrumentId, (Price, Price)>,
-    /// Last trade prices
-    pub last_trades: hashbrown::HashMap<InstrumentId, Price>,
-    /// Order counts by level
-    pub level_counts: hashbrown::HashMap<InstrumentId, (u32, u32)>, // (bid_count, ask_count)
-    /// Last update timestamps
-    pub last_updates: hashbrown::HashMap<InstrumentId, u64>,
-}
-
-impl FastMarketState {
-    pub fn new() -> Self {
-        Self {
-            best_prices: hashbrown::HashMap::new(),
-            last_trades: hashbrown::HashMap::new(),
-            level_counts: hashbrown::HashMap::new(),
-            last_updates: hashbrown::HashMap::new(),
-        }
-    }
-    
-    /// Update from market update (O(1) operations)
-    #[inline(always)]
-    pub fn update(&mut self, update: &MarketUpdate) {
-        match update {
-            MarketUpdate::Trade(trade) => {
-                self.last_trades.insert(trade.instrument_id, trade.price);
-                self.last_updates.insert(trade.instrument_id, trade.timestamp);
-            }
-            MarketUpdate::OrderBook(book_update) => {
-                self.last_updates.insert(book_update.instrument_id, book_update.timestamp);
-                // Note: For full BBO reconstruction, we'd need to maintain order books
-                // For now, we'll update on trades which is sufficient for many strategies
-            }
-        }
-    }
-    
-    /// Get best prices without locks
-    #[inline(always)]
-    pub fn get_best_prices(&self, instrument_id: InstrumentId) -> Option<(Price, Price)> {
-        self.best_prices.get(&instrument_id).copied()
-    }
-    
-    /// Get last trade price
-    #[inline(always)]
-    pub fn get_last_price(&self, instrument_id: InstrumentId) -> Option<Price> {
-        self.last_trades.get(&instrument_id).copied()
-    }
-    
-    /// Calculate spread efficiently
-    #[inline(always)]
-    pub fn get_spread(&self, instrument_id: InstrumentId) -> Option<i64> {
-        self.best_prices.get(&instrument_id)
-            .map(|(bid, ask)| ask.0 - bid.0)
-    }
-}
-
-/// Pre-allocated feature array with compile-time indices
-#[derive(Debug)]
-pub struct FastFeatureVector {
-    /// Values indexed by constants (no HashMap)
-    pub values: [f64; FEATURE_COUNT],
-    /// Timestamp
-    pub timestamp: u64,
-}
-
-// Pre-defined feature indices (zero-cost)
-pub const LAST_PRICE_IDX: usize = 0;
-pub const SPREAD_IDX: usize = 1;
-pub const PRICE_RETURN_IDX: usize = 2;
-pub const VOLUME_IDX: usize = 3;
-pub const TRADE_COUNT_IDX: usize = 4;
-pub const FEATURE_COUNT: usize = 5;
-
-impl FastFeatureVector {
-    pub fn new(timestamp: u64) -> Self {
-        Self {
-            values: [0.0; FEATURE_COUNT],
-            timestamp,
-        }
-    }
-    
-    /// Set feature by index (fastest possible)
-    #[inline(always)]
-    pub fn set(&mut self, index: usize, value: f64) {
-        if index < FEATURE_COUNT {
-            self.values[index] = value;
-        }
-    }
-    
-    /// Get feature by index
-    #[inline(always)]
-    pub fn get(&self, index: usize) -> f64 {
-        self.values.get(index).copied().unwrap_or(0.0)
-    }
-    
-    /// Update from market state (optimized)
-    #[inline(always)]
-    pub fn update_from_trade(&mut self, trade: &Trade, prev_price: Option<Price>) {
-        self.timestamp = trade.timestamp;
-        self.set(LAST_PRICE_IDX, trade.price.as_f64());
-        self.set(VOLUME_IDX, trade.quantity.as_f64());
-        
-        // Calculate return if we have previous price
-        if let Some(prev) = prev_price {
-            let ret = (trade.price.as_f64() - prev.as_f64()) / prev.as_f64();
-            self.set(PRICE_RETURN_IDX, ret);
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    
-    #[test]
-    fn test_fast_market_state() {
-        let mut state = FastMarketState::new();
-        
-        let trade = Trade {
-            instrument_id: 1,
-            price: Price::new(100_000_000),
-            quantity: Quantity::from(1000u32),
-            side: Side::Bid,
-            timestamp: 1000,
-        };
-        
-        state.update(&MarketUpdate::Trade(trade));
-        
-        assert_eq!(state.get_last_price(1), Some(Price::new(100_000_000)));
-    }
-    
-    #[test]
-    fn test_fast_features() {
-        let mut features = FastFeatureVector::new(1000);
-        
-        features.set(LAST_PRICE_IDX, 100.0);
-        features.set(VOLUME_IDX, 1000.0);
-        
-        assert_eq!(features.get(LAST_PRICE_IDX), 100.0);
-        assert_eq!(features.get(VOLUME_IDX), 1000.0);
-    }
-    
-    #[test]
-    fn test_side_conversion() {
-        assert_eq!(convert_side_i8(1), Side::Bid);
-        assert_eq!(convert_side_i8(2), Side::Ask);
-        assert_eq!(convert_side_i8(0), Side::Bid); // Default
-    }
-}
