+++ src/strategies/mod.rs
@@ -6,6 +6,7 @@
 pub mod market_maker;
 pub mod mean_reversion;
 pub mod trend_following;
+pub mod utils;
 
 pub use market_maker::MarketMakerStrategy;
 pub use mean_reversion::MeanReversionStrategy;
